[Android] OverviewModeObserver interface was refactored II
Chromium change:
https://chromium.googlesource.com/chromium/src/+/312e51357777601848825e0944165bc8e1fa4aba

Refactor the OverviewModeObserver interface

This CL mainly focus on moving the onOverviewModeStateChanged() signal
out of the OverviewModeObserver interface. That signal only matters when
StartSurface is enabled, so the signal is merged to
StartSurface.StateObserver.onStateChanged.

With the above change, this CL allows features to attach the
StartSurface.StateObserver to StartSurface via the ObservableSupplier
pattern.

In addition, this CL renames OverviewModeState to StartSurfaceState.

Change-Id: Iebaa907fca4baf6a1fee819b31fbd57370126dbc
Bug: 1108496