Revert Experiment from Canary that breaks Brave on Linux
This patch is part of an experiment (see crbug.com/1121427) and breaks
Chromium -and not just Brave- in a way that causes Chromium's top
level window to remain all-white on startup, with no widgets nor
content rendered at any point.

This patch has been reverted upstream already (along with other related
experiments) but we were unlucky on that 88.0.4298.5 had it applied but
not reverted yet, so we need to do it ourselves.

Please drop this patch once we rebase into a future version.

commit 78d26d83124f10a2b427e78d731afbf3e8e91a6b
Author: Bartek Nowierski <bartekn@chromium.org>
Date:   Wed Oct 21 22:19:04 2020 +0000

    Reland "Experiment on Canary: Enable PA-Everywhere (2nd round)"

    This is a reland of 252856b53f1dd91cd3b8f85a032f8215fd21c30e

    Original change's description:
    > Experiment on Canary: Enable PA-Everywhere (2nd round)
    >
    > Enables PartitionAlloc-Everywhere on Canary channel experimentally.
    >
    > Bug: 1121427
    > Change-Id: Icf6d473bb1cf1f1028be8249dd3e03cd369f6b16
    > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2490509
    > Reviewed-by: Krishna Govind <govind@chromium.org>
    > Cr-Commit-Position: refs/branch-heads/4298@{#4}
    > Cr-Branched-From: 2e88d10d4d625369d864ac4da7821767794772e6-refs/heads/master@{#818974}

    Bug: 1121427