Don't disable kPasswordCheck feature flag (it has been removed).
It is no longer possible to manually disable this feature but we don't
need to do it either since it only works for browsers allowing users
to sign-in (at the browser level i.e. via the IdentityManager) into
their Google accounts. Brave does not allow that, so there's nothing
that we need to disable and we can simply drop those 2 lines.

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/2c6129b54039cf821a97668d221fc59bb830fe7a

commit 2c6129b54039cf821a97668d221fc59bb830fe7a
Author: Jan Wilken Dörrie <jdoerrie@chromium.org>
Date:   Wed Dec 9 12:45:53 2020 +0000

    [Passwords] Remove PasswordCheck Feature Flag

    This change removes the PasswordCheck base::Feature flag. The feature
    was completely launched in M86.

    TBR=harrisonsean

    Bug: 1047726, 1075494, 1092444