Fixed unit tests These tests crashed on checking cryptographer_: C  391.819s Main  [  FAILED  ] BraveModelTypeWorkerTest.ResetProgressMarker (CRASHED) C  391.819s Main  [  FAILED  ] BraveModelTypeWorkerTest.ResetProgressMarkerDisabledFeature (CRASHED) C  391.819s Main  [  FAILED  ] BraveModelTypeWorkerTest.ResetProgressMarkerMaxPeriod (CRASHED)
Chromium change:
https://chromium.googlesource.com/chromium/src/+/79ebecad6367176e86a9ce3ddc6cedec9c3cfab9

[sync] Try to decrypt updates even if data type not listed as encrypted

Before this CL, Cryptographer was only injected in ModelTypeWorker if
the ModelType was listed as "encrypted". This meant if the worker of a
non-encrypted type unexpectedly received an encrypted update, it
couldn't decrypt the data even if the key *was* actually present in
the client. We have metrics showing some clients are in this 'powerless'
situation, so this CL changes the behavior. The cryptographer is now
always injected, and the worker always tries to decrypt updates. Commits
are still only encrypted if the type is listed as "encrypted", which is
known by the worker via a boolean.

If the unexpected encrypted update is decrypted, an unencrypted version
of the data should ideally be re-committed. This WAI for the incremental
update case, and the CL adds an integration test for this case. For
the initial sync case, code changes are needed and this is tracked in
crbug.com/1200321.

Other related changes:
- OnCryptographerChange() now called for non-encrypted types too.
- Sync.ModelTypeBlockedDueToUndecryptableUpdate.SomeKeysAvailable
recording removed and expiry date set to M92, since the worker won't
be blocked in this state anymore.
- EnableEncryption() now only calls UpdateTypeEncryptionKeyName()
instead of the whole OnCryptographerChange(). Because the worker is
always trying to decrypt anyways, setting enabled_encryption_ can't
suddenly unblock the worker and make it worth calling
DecryptStoredEntities() or NudgeIfReadyToCommit().

Somewhat non-related changes:
- Prevent changing the encryption key name in ModelTypeState from
non-empty to empty when the cryptographer has no default key. This
doesn't cause bugs today because BlockForEncryption() is true in this
case, so the processor never sees the empty string. But doesn't hurt to
make the code more robust. Also rename the corresponding method to make
it more obvious that it refers to the key name in ModelTypeState.
- Move tests for PopulateUpdateResponseData() to a separate fixture-less
suite, since the method is static and requires none of the worker utils
in ModelTypeWorkerTest.

Bug: 1178418, 1109221