Remove dependency on no longer existing target //base/util/timer:timer
This was initially needed as per some changes that were going to be
added as part of PR #8124[1] (in commit 8ba9ddb5[2]), that used
util::WallClockTimer, but the merged PR ended up implementing things
differently and this dependency was left around in the BUILD.gn file.

Since //base/util/timer does not exist anymore in Chromium 94 (i.e it
has been moved to //base/timer), this is breaking the build now so
let's remove the stray dependency.

[1] https://github.com/brave/brave-core/pull/8124
[2] https://github.com/brave/brave-core/commit/8ba9ddb5