Fix build error due to non-existing types after migration to TypeScript
Brave started failing building on Chromium 96.0.4651.0 with this:

  RuntimeError: Command '/home/mario/work/brave-rebases/brave-browser-next/src/third_party/node/linux/node-linux-x64/bin/node /home/mario/work/brave-rebases/brave-browser-next/src/third_party/node/node_modules/typescript/bin/tsc --project gen/chrome/browser/resources/settings/tsconfig.json' failed
  gen/chrome/browser/resources/settings/preprocessed/site_settings/category_default_setting.ts:201:33 - error TS2339: Property 'AUTOPLAY' does not exist on type 'typeof ContentSettingsTypes'.
  201       case ContentSettingsTypes.AUTOPLAY:
                                    ~~~~~~~~
  gen/chrome/browser/resources/settings/preprocessed/site_settings/category_default_setting.ts:236:33 - error TS2339: Property 'ETHEREUM' does not exist on type 'typeof ContentSettingsTypes'.
  236       case ContentSettingsTypes.ETHEREUM:
                                    ~~~~~~~~
It seems that the problem would be that the migration to TypeScript of
category_default_setting.js means that we can no longer rely on the way
resources/settings/brave_overrides/index.js was used to extend the
ContentSettingsType JS enumeration, leading to that build error.

Therefore, we have to go back to patching the constants.ts file as
we did in the past (see change in [1]), to extend the enumeration.

[1] https://github.com/brave/brave-core/commit/09f2d021af4