[Android] Java code adjustments
Chromium changes:
https://github.com/chromium/chromium/commit/670a9e15d126af5600808cc08733447d5d474cd9

Update app menu to use MVC framework
Update AppMenu to use ModelListAdapter, also pass data between M-V-C by
ModelList instead of Menu.

BUG:1169999

https://github.com/chromium/chromium/commit/dd18c443f32d30a25395f07c82b5cc0680ec54eb

[AutoDark] Add Android auto dark metrics
Adding metrics that tracks the source for auto dark settings changes,
and theme settings entries.

Noting some of the enums are not used yet in this CL. There are added
based on the design doc (See http://shortn/_enU3FSpUBa)

Bug: 1241542, 1250800

https://github.com/chromium/chromium/commit/c6942d6678711188bfaf14ac85c9c6ee6c87be97

[MerchantTrust] Add store icon and related IPH to omnibox
This CL adds the store icon and its related IPH to omnibox. The icon is
shown when the merchant trust message is dismissed. We will add the icon
metrics and the IPH config in the later CLs.

Bug: 1252543