Adds an override for insecure variations server URL.
Network audit test error:

NETWORK AUDIT FAIL:http://clientservices.googleapis.com/chrome-variations/seed?osname=linux&milestone=98

While we already override the secure variations server URL, if the
connection doesn't succeed for whatever reason, VariationsServer may try
to use the fall-back insecure url, which we weren't overriding.

We don't want to fall back on an insecure url, though. So, instead we
supply the same secure url to the insecure url switch. This will cause
the code in the VariationsServer to not use it.