Overriding token limit in content_browser_client.cc
As explained in the comment above, this token limit has been introduced
to this translation unit to prevent it from gettting too large. However,
brave's header override for that file is now hitting the limit, and
since the file we are adding is a mojom header file, there's not much
that can be done through the use of forward declarations.

This change slightly increases the token limit to accomodate for the
necessary brave overrides inserted into the header.