Changes for BasicStartupComplete
Chromium change:
https://github.com/chromium/chromium/commit/3b0d75acbde026dfefa4cced40b92ac53bc19281

Add a shortcut path to exit from ContentMainDelegate APIs
This CL is adding a shortcut path to allow embedder to early exit
the process just before it becomes the browser process. The
functions PreBrowserMain(...) and PostEarlyInitialization(...)
can early exit the startup.

This CL is a no-op and should have no functional change. It's boiler
plate to allow the refactoring of the Chrome process singleton.

The goal of this change is to allow to execute the process singleton
(which guarantee only one browser process is running) much earlier
in the Chrome startup sequence.

Bug: 1340599