[Android] New method for DropdownItemProcessor
Chromium change:
https://github.com/chromium/chromium/commit/da78d8f66a8411a39e67285fc1bdb280f7ed0def

[Omnibox] Implement background for the suggestions
screenshot:
https://screenshot.googleplex.com/8fx7B8kkSroZAag

This CL adds background drawable to the Omnibox suggestions, and adds
space between suggestions.
This CL also adds the flag OMNIBOX_MODERNIZE_VISUAL_UPDATE to be cached
so search activity can also uses it when native is not loaded.

This CL only updates the suggestions' UI, not Omnibox. The CL for
Omnibox will be done later.

Bug:1348445