[Android] Various fixes to adjust java code to upstream changes
Chromium changes:
https://github.com/chromium/chromium/commit/f002437bf1838e0915129ccff21bf52a37577a41

[Omnibox] Prevent LocationBar crash when onMeasured entered post dest…
…roy on ToolbarPhone

Bug: 1359257

https://github.com/chromium/chromium/commit/df8db0da2e09b4cae37fa15733a17105b8b568d7

[WebContentsImpl] Don't call didFinishNavigation in child frames
Part of reducing navigation cost to reduce scroll jank. Java observers
don't do work when didFinishNavigation is called in child frames,
so we don't have to notify java in this case.

The changes to restrict didFinishNavigation to main frames is not behind a flag, so to measure the effect on jank, we added a didFinishNavigationNoop function (behind a flag) to mirror what didFinishNavigation did before this change. We think that the overhead from overrides add up to something measurable, so I added a redundant boolean check to make sure this override doesn't get optimized (might get optimized if it's empty but not completely sure). I also wanted it to be as close as possible to what it was doing before, so that we can capture more accurately the effect on jank metrics.

Bug: 1359110

https://github.com/chromium/chromium/commit/a603bf1de79b2a844bb05c6252e5c1e97938f591

Suppress captures in CCT toolbar.
CustomTabToolbar maintains a token object to compare states. Only
allow captures when the previously captured token is different from
the current token.

Bug: 1363389

https://github.com/chromium/chromium/commit/1900d0973093225f3cc844693841b4f15b970a4f

[Start] Add a logo in Start surface below toolbar.
This CL adds a big logo in Start surface (similar to the logo on NTP) when Feed is not shown.
Video: http://shortn/_5DZDGESzIU
Screenshot: https://screenshot.googleplex.com/AbKoKQJy4qPtYVp

The follow-up CLs would be :
1. Fix the omnibox scrolling animation
2. Replace logoView in tasks_view_layout with ViewStub

Bug: 1365694