[Android] New c-tor signature for TopToolbarCoordinator
Chromium change:
https://github.com/chromium/chromium/commit/65c903d7aa0c4500b8cbfe475951a567125700d8

Suppress top toolbar captures on touch.
Plumbs touch/scroll events from CompositorViewHolder to the top
toolbar adapter. This captures to be blocked during scroll. In order
to prevent old captures from being shown during this window, on demand
controls locking is used. This must be done as fast as possible, when
the touch starts, not when the render frame is being created,
otherwise artifacts are visible.

(cherry picked from commit 6136f92)

Bug: 1374427