[Android] Refactored notification rationale dialog as a bottom sheet
Chromium change:
Refactored notification rationale dialog as a bottom sheet
Android T+ requires apps to request for notification permission.
Chrome does this currently on first startup, and it shows a rationale
dialog before the OS prompt to educate users of the usefulness of
notifications.

Per UX advice this UI should be displayed as a bottom sheet instead of
a dialog. This CL introduces a bottom sheet that contains the same
UI as the existing rationale dialog. It's behind a flag exposed on
chrome://flags.

It also removes an alternative set of strings that were used when experimenting with this UI.

Bottomsheet screenshot:
https://screenshot.googleplex.com/84j6m6brZBz6uSp
HSV:
https://hsv.googleplex.com/6181362051579904

Bug: 1361424