[Android] Autocomplete setting was renamed
Chromium change:
https://github.com/chromium/chromium/commit/e152def0ed5414f72091d1d784f609f2ac646704

[omnibox] Rename omnibox autocomplete setting
Old:
Title - "Autocomplete searches and URLs"
Description - "Sends some cookies and searches from the address bar and search box to your default search engine"

New:
Title - "Improve search suggestions"
Description - "When you type in the address bar or search box, Chrome sends what you type to your default search engine to get better suggestions. This is off in Incognito."

Additionally, moves the setting lower in the list of settings in the UI so that on desktop it's next to the Drive suggest setting and on Android and iOS it's consistent with the new position on desktop.

Bug: 1179549