[Android] Rename onUrlFocusChange to onOmniboxSessionStateChange
Chromium change:
https://github.com/brave/chromium/commit/698cb6c47e3e1005941f5b4e0289ba594454a4c7

[APS] Trigger zero-prefix dropdown on tapping the omnibox after it is
auto-focused

Opening the NTP automatically focuses on the omnibox with a blinking
cursor and without triggering the zero-prefix dropdown when a physical
keyboard is connected to a tablet device.

This CL adds support to show the zero-prefix dropdown when the omnibox
is tapped at this point.

This CL also renames/repurposes AutocompleteMediator#onUrlFocusChange() to report metrics and activate/deactivate the autocomplete session as follows:
1. The autocomplete session will not be activated when the omnibox is initially focused. This will be done only if the user taps on / types into the omnibox after such focus.
2. If the user chooses not to interact with the omnibox after such focus, the autocomplete session deactivation logic in this method will not be executed.

Bug: 1445822