[Android] Adjustments for DropdownItemViewInfoListBuilder changes
Chromium change:
https://github.com/brave/chromium/commit/ddfa614c3740b47df416e7a944e4086f281c1bae

Relocate query extraction for SRO Query Edit to Native code.
The change is functionally a no-op.

Note: the change drops verification of validity of destination url in
native code. This is because the mechanism used to classify matches
has been replaced over the last few years with one that no longer
rewrites entire AutocompleteMatch.

Change-Id: I815dabcf16bca0cdffd8755f6aa54da17ad7e7a4