Delete used `CreateVectorValuesPtr`
This function overload was causing a build error as it is unused. The
correct solution here would be to have template specialisations,
however to simplify the removal of the unused overload, this change
merely deletes it, and adds a `static_assert`, to avoid accidental
assumptions about its presence in future changes.