Fix `BraveBrowsingDataRemoverDelegate` substitution
This override has been missing a matching header since it was
introduced, and this has been causing linking to fail in every upstream
occurrence of `ChromeBrowsingDataRemoverDelegateFactory::GetForProfile`
as the return type in the header was not being replaced accordingly.

This changes adds the changes to the header as well, and provides an
inclusion `BraveBrowsingDataRemoverDelegate` in the shadow file, to
allow the compiler to understand that
`BraveBrowsingDataRemoverDelegate*` is pefectly convertible to
`ChromeBrowsingDataRemoverDelegate`.