`StartupBrowserCreator` now takes a tab restore arg
PWAs do not need to restore a normal browser, therefore upstream is
providing an extra argument for `Launch` to indicate that a normal store
doesn't need to occur.

In brave this function is overriden, and the override is being corrected
accordingly to match the change in upstream. In all the other places
this function is being called, we are now passing the arg as true to
retore the browser, as these are browser tests.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/34ff0ccb88702c1d62c63706d00980dc7a3dd2b9

commit 34ff0ccb88702c1d62c63706d00980dc7a3dd2b9
Author: Pengchao Cai <pengchaocai@chromium.org>
Date:   Wed Feb 7 03:35:46 2024 +0000

    [lacros] Stop restoring a normal browser when restoring only PWAs

    Currently a normal browser is restored by default even with only PWAs
    open in previous session. This CL introduces a bool parameter on session
    restore code path and it only gets flipped to false if invoked by ash
    full restore, indicating that if this is a full restore, don't create
    a normal browser by default.

    This does not affect cases when there were normal browser windows open
    in previous session.

    Change-Id: I98a3abfa4ddc0da6be6aa613181f50047a43872b
    Bug: 1463906