[Android] Update Read Aloud magic toolbar button when page is translated
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/ee6035ab3770ea4e92fead9dd0622cfab6d75388

Update Read Aloud magic toolbar button when page is translated

The page's readability may have changed if it was translated, so the
button should check again and possibly show or hide itself.

Underlying changes:
- Add a TranslationObserver to the selected tab as needed
- Change how ReadAloudController notifies outside features of readability events

This is a reland of https://chromium-review.googlesource.com/c/chromium/src/+/5319377 with fixes: the TranslationObserver can't be registered in onPageLoadStarted() because it runs for a temporary WebContents during tab restore and there's no chance to unregister.

Bug: 326287049