Fixes CommandUtilsBrowserTest timing out.
The timeout was likely due to executing Add new profile command which would
block on the profile name dialog and all consequent commands would be
attempted to be executed on that browser. Added code to select the first
browser each time before executing the command.

Also, on Windows and Linux, the test was likely exiting early because
IDC_CLOSE_TAB and IDC_CLOSE_WINDOW would exit the browser if they were
the only tab/window present.