[rust] Use `byteorder` crate checked out in chromium
This crate is used in brave, and it has now been introduced in upstream
chromium as well. This change deletes the local copy for this crate, as
this was causing duplicated id errors for the two crates, and makes sure
we use the crate checked out on upstream. However, some feature patching
was required for the use we do in brave.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/6cc428454a87621f6cefbc08050ca53ded4cfca7

commit 6cc428454a87621f6cefbc08050ca53ded4cfca7
Author: danakj <danakj@chromium.org>
Date:   Fri Sep 13 20:46:38 2024 +0000

    Import zerocopy and byteorder crates

    The zerocopy crate is used by rand now (via ppv-lite86) so we pull it in
    to be able to roll our 3p dependencies to their latest versions. As it
    is developed by and used internally at Google, we'll see future crates
    make use of it as well.

    The zerocopy crate provides safe abstractions to handle data in a
    zero-copy manner, which we would grately prefer over writing unsafe Rust
    code, so we allow the crate for 1p use while importing it.

    Unsafe review for byteorder is done in this CL.

    Change-Id: I077b0596bd3738ecd26da8ad2aab168626f23d2f