[Android] Rename secure state icons
Chromium change:
https://chromium.googlesource.com/chromium/src/+/18371dfa10268ba754ca5cbd75beda9038f82503

[omnibox] Rename Android secure state icons to reflect their function.

The omnibox secure state indicator was refreshed last year to show a page info icon instead of a lock icon. As of https://crrev.com/c/5982459, the refreshed state became the only state for the omnibox indicator but the icon was still named "refresh" and the old lock icon was still in use in other surfaces.

This CL changes the icon names to "page_info" and "lock" to clearly distinguish them.

Change-Id: Ic7175efccbed0dbbd1523827873b26c4209b4e67