[Android] Update Android secure connections setting
Chromium change:
https://chromium.googlesource.com/chromium/src/+/0508b9601fc32616c643353e68e60df750a54710

[HFM] Update Android secure connections setting

This adds a new settings UI for HTTPS-First Mode ("Always use secure
connections") to have a separate sub-page with the overall feature
toggle and a sub-section radio group for selecting between Balanced
Mode and Strict Mode. The new settings layout, logic, and tests are
in a new chrome/browser/ssl/android/ sub-directory.

This is currently gated on the HttpsFirstBalancedMode feature flag.

This also moves the logic for updating the underlying prefs based on
the UI setting to a method on HttpsFirstModeService so that it can
be reused between the Desktop WebUI settings code and the Android
Java settings code.

The existing tests for the old Android settings UI are renamed to be
"legacy" and can be removed once the new settings UI is fully rolled
out.

NO_IFTTT=Adding generated Java enum only, no change needed to enums.xml

Bug: 349860796