fixup! [Android] Fix for login DB deprecation export service
We intend to use LoginDatabase as the backend for password manager going forward, so here we just explicitly call Chromium's code in else block to avoid error:

../../chrome/browser/password_manager/android/login_db_deprecation_runner_factory.cc:43:22: error: code will never be executed [-Werror,-Wunreachable-code]
   43 |   Profile* profile = Profile::FromBrowserContext(context);

Chromium change:
https://chromium.googlesource.com/chromium/src/+/193e77705d4c5176af6571407fdfe753bf215f68

[UPMLoginDbSunset] Add factory for DB deprecation export service

It owns a service which creates the exporter and destroys it once
the export is done. It only creates the service for clients who have unmigrated passwords.

Bug: 378650395