[Android] Disable quick delete flag
When it is enabled we have missing Brave's items in `Delete browsing data` option. This issue sill be addresse in C134 when flag is completely removed.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/aec661770742e09205cfcdcbf03dbd1ac8cf777b

Enable QD Follow up on ToT

We are proceeding with full launch on the experiment arm that will
open a new tab when all tabs are deleted via DBD.
This change will clean up other arms and leaves the main arm as the
default behavior.

Launch: launch/4307414