[cr140][Android] WIP PersonalDataManager.getProfilesToSuggest has no arguments anymore
Marked as WIP because we were using includeNameInLabel=true, but upstream used
includeNameInLabel=false, so it is defaulted to false.

these lines at PersonalDataManagerAndroid::GetProfileLabels
    if (include_name_in_label) {
      suggested_fields.insert(NAME_FULL);
    }
are not executed anymore.
Instead /*excluded_fields=*/{NAME_FULL} is now passed into
AutofillProfile::CreateInferredLabels. We need to figure out if it is safe to leave it as is.

Chromium change:
https://source.chromium.org/chromium/chromium/src/+/8bed5e34c6175679c2c5250dea84ef49b9531a39

	[Android] Simplify PersonalDataManager JNI calls

	This CL simplifies the `getProfilesToSuggest()`,
	`getProfileLabelsToSuggest()`, and `GetProfileLabels()` methods by
	inlining their constant arguments.

	Bug: 40100455
	Change-Id: I4ef7cf9286d849dfd2b7699987a45ac6caab769b
	Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6771813