[cr145] Use AsyncMemoryPressureListenerRegistration for ads/rewards dbs.
The non-async registration now uses a CHECK in upstream code to
verify that the registration is done on the main thread.

Per Chromium comments, async version is used for listeners that
can exists on sequences other than the main thread and don't need
to be called synchronously. This seems to fit our use cases.

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/facfee34d5daf6bbde99f4c5a63f3a6bf229e72b

commit facfee34d5daf6bbde99f4c5a63f3a6bf229e72b
Author: Patrick Monette <pmonette@chromium.org>
Date:   Mon Nov 10 19:22:01 2025 -0800

    Ensure that NotifyMemoryPressure is called on the main thread

    Bug: 457269074